+++ cmd/controller-manager/main.go
@@ -22,7 +22,7 @@ import (
 
 	chaosmeshv1alpha1 "github.com/chaos-mesh/chaos-mesh/api/v1alpha1"
 	apiWebhook "github.com/chaos-mesh/chaos-mesh/api/webhook"
-	"github.com/chaos-mesh/chaos-mesh/controllers/common"
+	ccfg "github.com/chaos-mesh/chaos-mesh/controllers/config"
 	"github.com/chaos-mesh/chaos-mesh/controllers/metrics"
 	"github.com/chaos-mesh/chaos-mesh/controllers/podiochaos"
 	"github.com/chaos-mesh/chaos-mesh/controllers/podnetworkchaos"
@@ -89,23 +89,23 @@ func main() {
 	}
 
 	// set RPCTimeout config
-	utils.RPCTimeout = common.ControllerCfg.RPCTimeout
+	utils.RPCTimeout = ccfg.ControllerCfg.RPCTimeout
 
 	ctrl.SetLogger(zap.Logger(true))
 
 	options := ctrl.Options{
 		Scheme:             scheme,
-		MetricsBindAddress: common.ControllerCfg.MetricsAddr,
-		LeaderElection:     common.ControllerCfg.EnableLeaderElection,
+		MetricsBindAddress: ccfg.ControllerCfg.MetricsAddr,
+		LeaderElection:     ccfg.ControllerCfg.EnableLeaderElection,
 		Port:               9443,
 	}
 
-	if common.ControllerCfg.ClusterScoped {
+	if ccfg.ControllerCfg.ClusterScoped {
 		setupLog.Info("Chaos controller manager is running in cluster scoped mode.")
 		// will not specific a certain namespace
 	} else {
-		setupLog.Info("Chaos controller manager is running in namespace scoped mode.", "targetNamespace", common.ControllerCfg.TargetNamespace)
-		options.Namespace = common.ControllerCfg.TargetNamespace
+		setupLog.Info("Chaos controller manager is running in namespace scoped mode.", "targetNamespace", ccfg.ControllerCfg.TargetNamespace)
+		options.Namespace = ccfg.ControllerCfg.TargetNamespace
 	}
 
 	mgr, err := ctrl.NewManager(ctrl.GetConfigOrDie(), options)
@@ -114,7 +114,7 @@ func main() {
 		os.Exit(1)
 	}
 
-	err = router.SetupWithManagerAndConfigs(mgr, common.ControllerCfg)
+	err = router.SetupWithManagerAndConfigs(mgr, ccfg.ControllerCfg)
 	if err != nil {
 		setupLog.Error(err, "fail to setup with manager")
 		os.Exit(1)
@@ -137,7 +137,7 @@ func main() {
 		Client:                  mgr.GetClient(),
 		Reader:                  mgr.GetAPIReader(),
 		Log:                     ctrl.Log.WithName("handler").WithName("PodNetworkChaos"),
-		AllowHostNetworkTesting: common.ControllerCfg.AllowHostNetworkTesting,
+		AllowHostNetworkTesting: ccfg.ControllerCfg.AllowHostNetworkTesting,
 	})
 	if err = (&chaosmeshv1alpha1.PodNetworkChaos{}).SetupWebhookWithManager(mgr); err != nil {
 		setupLog.Error(err, "unable to create webhook", "webhook", "PodNetworkChaos")
@@ -149,25 +149,25 @@ func main() {
 
 	setupLog.Info("Setting up webhook server")
 	hookServer := mgr.GetWebhookServer()
-	hookServer.CertDir = common.ControllerCfg.CertsDir
+	hookServer.CertDir = ccfg.ControllerCfg.CertsDir
 	conf := config.NewConfigWatcherConf()
 
 	stopCh := ctrl.SetupSignalHandler()
 
-	if common.ControllerCfg.PprofAddr != "0" {
+	if ccfg.ControllerCfg.PprofAddr != "0" {
 		go func() {
-			if err := http.ListenAndServe(common.ControllerCfg.PprofAddr, nil); err != nil {
+			if err := http.ListenAndServe(ccfg.ControllerCfg.PprofAddr, nil); err != nil {
 				setupLog.Error(err, "unable to start pprof server")
 				os.Exit(1)
 			}
 		}()
 	}
 
-	if err = common.ControllerCfg.WatcherConfig.Verify(); err != nil {
+	if err = ccfg.ControllerCfg.WatcherConfig.Verify(); err != nil {
 		setupLog.Error(err, "invalid environment configuration")
 		os.Exit(1)
 	}
-	configWatcher, err := watcher.New(*common.ControllerCfg.WatcherConfig, metricsCollector)
+	configWatcher, err := watcher.New(*ccfg.ControllerCfg.WatcherConfig, metricsCollector)
 	if err != nil {
 		setupLog.Error(err, "unable to create config watcher")
 		os.Exit(1)
@@ -177,7 +177,7 @@ func main() {
 	hookServer.Register("/inject-v1-pod", &webhook.Admission{
 		Handler: &apiWebhook.PodInjector{
 			Config:        conf,
-			ControllerCfg: common.ControllerCfg,
+			ControllerCfg: ccfg.ControllerCfg,
 			Metrics:       metricsCollector,
 		}},
 	)
