+++ controllers/common/types.go
@@ -15,22 +15,16 @@ package common
 
 import (
 	"context"
-	"fmt"
-	"os"
-	"strings"
 	"time"
 
 	"github.com/chaos-mesh/chaos-mesh/api/v1alpha1"
-	"github.com/chaos-mesh/chaos-mesh/pkg/config"
 	ctx "github.com/chaos-mesh/chaos-mesh/pkg/router/context"
 	endpoint "github.com/chaos-mesh/chaos-mesh/pkg/router/endpoint"
-	"github.com/chaos-mesh/chaos-mesh/pkg/utils"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/util/retry"
 
 	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 )
 
 const (
@@ -43,53 +37,6 @@ const (
 const emptyString = ""
 
 var log = ctrl.Log.WithName("controller")
-
-// ControllerCfg is a global variable to keep the configuration for Chaos Controller
-var ControllerCfg *config.ChaosControllerConfig
-
-func init() {
-	conf, err := config.EnvironChaosController()
-	if err != nil {
-		ctrl.SetLogger(zap.Logger(true))
-		log.Error(err, "Chaos Controller: invalid environment configuration")
-		os.Exit(1)
-	}
-
-	err = validate(&conf)
-	if err != nil {
-		ctrl.SetLogger(zap.Logger(true))
-		log.Error(err, "Chaos Controller: invalid configuration")
-		os.Exit(1)
-	}
-
-	ControllerCfg = &conf
-}
-
-func validate(config *config.ChaosControllerConfig) error {
-
-	if config.WatcherConfig == nil {
-		return fmt.Errorf("required WatcherConfig is missing")
-	}
-
-	if config.ClusterScoped != config.WatcherConfig.ClusterScoped {
-		return fmt.Errorf("K8sConfigMapWatcher config ClusterScoped is not same with controller-manager ClusterScoped. k8s configmap watcher: %t, controller manager: %t", config.WatcherConfig.ClusterScoped, config.ClusterScoped)
-	}
-
-	if !config.ClusterScoped {
-		if strings.TrimSpace(config.TargetNamespace) == "" {
-			return fmt.Errorf("no target namespace specified with namespace scoped mode")
-		}
-		if !utils.IsAllowedNamespaces(config.TargetNamespace, config.AllowedNamespaces, config.IgnoredNamespaces) {
-			return fmt.Errorf("target namespace %s is not allowed with filter, please check config AllowedNamespaces and IgnoredNamespaces", config.TargetNamespace)
-		}
-
-		if config.TargetNamespace != config.WatcherConfig.TargetNamespace {
-			return fmt.Errorf("K8sConfigMapWatcher config TargertNamespace is not same with controller-manager TargetNamespace. k8s configmap watcher: %s, controller manager: %s", config.WatcherConfig.TargetNamespace, config.TargetNamespace)
-		}
-	}
-
-	return nil
-}
 
 // Reconciler for common chaos
 type Reconciler struct {
