+++ controllers/config/config.go
@@ -1,0 +1,97 @@
+// Copyright 2019 Chaos Mesh Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package config
+
+import (
+	"fmt"
+	"os"
+	"regexp"
+	"strings"
+
+	"sigs.k8s.io/controller-runtime/pkg/log/zap"
+
+	"github.com/chaos-mesh/chaos-mesh/pkg/config"
+
+	ctrl "sigs.k8s.io/controller-runtime"
+)
+
+// ControllerCfg is a global variable to keep the configuration for Chaos Controller
+var ControllerCfg *config.ChaosControllerConfig
+
+var log = ctrl.Log.WithName("config")
+
+func init() {
+	conf, err := config.EnvironChaosController()
+	if err != nil {
+		ctrl.SetLogger(zap.Logger(true))
+		log.Error(err, "Chaos Controller: invalid environment configuration")
+		os.Exit(1)
+	}
+
+	err = validate(&conf)
+	if err != nil {
+		ctrl.SetLogger(zap.Logger(true))
+		log.Error(err, "Chaos Controller: invalid configuration")
+		os.Exit(1)
+	}
+
+	ControllerCfg = &conf
+}
+
+func validate(config *config.ChaosControllerConfig) error {
+
+	if config.WatcherConfig == nil {
+		return fmt.Errorf("required WatcherConfig is missing")
+	}
+
+	if config.ClusterScoped != config.WatcherConfig.ClusterScoped {
+		return fmt.Errorf("K8sConfigMapWatcher config ClusterScoped is not same with controller-manager ClusterScoped. k8s configmap watcher: %t, controller manager: %t", config.WatcherConfig.ClusterScoped, config.ClusterScoped)
+	}
+
+	if !config.ClusterScoped {
+		if strings.TrimSpace(config.TargetNamespace) == "" {
+			return fmt.Errorf("no target namespace specified with namespace scoped mode")
+		}
+		if !IsAllowedNamespaces(config.TargetNamespace, config.AllowedNamespaces, config.IgnoredNamespaces) {
+			return fmt.Errorf("target namespace %s is not allowed with filter, please check config AllowedNamespaces and IgnoredNamespaces", config.TargetNamespace)
+		}
+
+		if config.TargetNamespace != config.WatcherConfig.TargetNamespace {
+			return fmt.Errorf("K8sConfigMapWatcher config TargertNamespace is not same with controller-manager TargetNamespace. k8s configmap watcher: %s, controller manager: %s", config.WatcherConfig.TargetNamespace, config.TargetNamespace)
+		}
+	}
+
+	return nil
+}
+
+// IsAllowedNamespaces returns whether namespace allows the execution of a chaos task
+func IsAllowedNamespaces(namespace string, allowedNamespaces, ignoredNamespaces string) bool {
+	if allowedNamespaces != "" {
+		matched, err := regexp.MatchString(allowedNamespaces, namespace)
+		if err != nil {
+			return false
+		}
+		return matched
+	}
+
+	if ignoredNamespaces != "" {
+		matched, err := regexp.MatchString(ignoredNamespaces, namespace)
+		if err != nil {
+			return false
+		}
+		return !matched
+	}
+
+	return true
+}
