+++ pkg/utils/selector_test.go
@@ -448,56 +448,6 @@ func TestFilterByAnnotations(t *testing.T) {
 	}
 }
 
-func TestIsAllowedNamespaces(t *testing.T) {
-	g := NewGomegaWithT(t)
-	type TestCase struct {
-		name   string
-		pods   []v1.Pod
-		ret    []bool
-		allow  string
-		ignore string
-	}
-	pods := []v1.Pod{
-		newPod("p1", v1.PodRunning, "allow", nil, nil, ""),
-		newPod("p1", v1.PodRunning, "allow-app", nil, nil, ""),
-		newPod("p1", v1.PodRunning, "app-allow", nil, nil, ""),
-		newPod("p1", v1.PodRunning, "ignore", nil, nil, ""),
-		newPod("p1", v1.PodRunning, "ignore-app", nil, nil, ""),
-		newPod("p1", v1.PodRunning, "app-ignore", nil, nil, ""),
-	}
-
-	allowRet := []bool{true, true, true, false, false, false}
-
-	var tcs []TestCase
-	tcs = append(tcs, TestCase{
-		name:  "only set allow",
-		pods:  pods,
-		ret:   allowRet,
-		allow: "allow",
-	})
-
-	tcs = append(tcs, TestCase{
-		name:   "only set ignore",
-		pods:   pods,
-		ret:    allowRet,
-		ignore: "ignore",
-	})
-
-	tcs = append(tcs, TestCase{
-		name:   "only set allow",
-		pods:   pods,
-		ret:    allowRet,
-		allow:  "allow",
-		ignore: "ignore",
-	})
-
-	for _, tc := range tcs {
-		for index, pod := range tc.pods {
-			g.Expect(IsAllowedNamespaces(pod.Namespace, tc.allow, tc.ignore)).Should(Equal(tc.ret[index]))
-		}
-	}
-}
-
 func TestFilterNamespaceSelector(t *testing.T) {
 	g := NewGomegaWithT(t)
