+++ pkg/webhook/inject/inject.go
@@ -24,6 +24,8 @@ import (
 	"github.com/chaos-mesh/chaos-mesh/pkg/utils"
 	"github.com/chaos-mesh/chaos-mesh/pkg/webhook/config"
 
+	ccfg "github.com/chaos-mesh/chaos-mesh/controllers/config"
+
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
@@ -142,7 +144,7 @@ func injectRequired(metadata *metav1.ObjectMeta, cli client.Client, cfg *config.
 			return "", false
 		}
 	}
-	if !utils.IsAllowedNamespaces(metadata.Namespace, controllerCfg.AllowedNamespaces, controllerCfg.IgnoredNamespaces) {
+	if !ccfg.IsAllowedNamespaces(metadata.Namespace, controllerCfg.AllowedNamespaces, controllerCfg.IgnoredNamespaces) {
 		log.Info("Skip mutation for it' in special namespace", "name", metadata.Name, "namespace", metadata.Namespace)
 		return "", false
 	}
